Modification in file structure : PromosController creation
