--- src/Unsapa/IPWBundle/Resources/views/Admin/add.html.twig
@@ -1,11 +1,0 @@
-{% extends "UnsapaIPWBundle::layout.html.twig" %}
-
-{% block content %}
-  {% include "UnsapaIPWBundle:Forms:promo.html.twig" %}
-{% endblock %}
-
-{# ???
-{% block footer_js %}
-<script type="text/javascript" src="{{ asset('bundles/unsapaipw/js/add_exam.js') }}"></script>
-{% endblock %}
-#}
