+++ src/Unsapa/IPWBundle/Resources/views/Forms/promo.html.twig
@@ -1,6 +1,6 @@
 <h1>Ajout d'une promotion</h1>
 
-<form action="{{ path('admin_add') }}" method="post" {{ form_enctype(form) }}>
+<form action="{{ path('promos_add') }}" method="post" {{ form_enctype(form) }}>
     {{ form_errors(form) }}
     {{ form_widget(form) }}
