+++ src/Unsapa/IPWBundle/Resources/views/Promos/add.html.twig
@@ -1,0 +1,5 @@
+{% extends "UnsapaIPWBundle::layout.html.twig" %}
+
+{% block content %}
+  {% include "UnsapaIPWBundle:Forms:promo.html.twig" %}
+{% endblock %}
