+++ src/Unsapa/IPWBundle/Resources/views/Promos/index.html.twig
@@ -1,0 +1,5 @@
+{% if is_granted("ROLE_ADMIN") %}
+  {% include "UnsapaIPWBundle:Promos:index_admin.html.twig" %}
+{% else %}
+  {% include "UnsapaIPWBundle:Promos:index_user.html.twig" %}
+{% endif %}
