+++ src/Unsapa/IPWBundle/Resources/views/Promos/index_user.html.twig
@@ -1,0 +1,13 @@
+{% extends 'UnsapaIPWBundle::layout.html.twig' %}
+ 
+{% block title %}Promotions{% endblock %}
+
+{% block content %}
+	
+	<ul>
+	{% for promo in promos %}
+		<li>{{ promo.name }}
+	{% endfor %}
+	</ul>
+	
+{% endblock %}
